Add Create Destination dialog
A new dialog that can be open using the context menu
on a file entry. It allows the user to create a new destination.
The file name is displayed there too. Allowing the user
to click on parts of the file name to add those words
to the new destination name.
And an add name button to add name separators
dynamically.