Fix race condition cancelling Thumbnail tasks
The thumbnail tasks in the FileComparison dialog could cause
a race condition when both threads were still running but when
the first finished, it called the close() function again.
This triggered the OnClose event, checking if any of the tasks
is still running. Which was true for the second one. But it
tried to cancel the first task too. Causing it to crash because
the CancellationTokenSource was already disposed.