Add highlight for better file
When source and destination have the same name
the files are compared for length and size.
If the length is different, the user has to decide.
Otherwise the file size is compared. If the new
file is at least 100 kB larger, it will highlight the
new file. Otherwise the old file will be highlighted.