Fix issue with open file
When trying to move a file that is opened, it caused an IOException.
Which was detected just as if the file already existed in the
destination. Which tried to open the compare dialog. This caused a
FileNotFoundException when trying to read the properties of the
destination file. Now, it checks that the IOException does have the
"already exists" string in the message to avoid this issue.